LayoutTests/imported/w3c/web-platform-tests/css/css-grid/abspos/grid-positioned-items-and-autofit-tracks-004-expected.txt


FAIL .grid 1 assert_equals: 
<div class="container">
    <div class="grid">
        <span style="grid-column: 1 / 5" class="abs" data-expected-width="30" data-expected-height="10"></span>
        <span style="grid-column: 1" data-expected-width="30" data-expected-height="10"></span>
    </div>
</div>
width expected 30 but got 25
